Refactor: Separated navbar component in shared components and modularized in styles, component, interfaces.
