+++ src/components/index.tsx
@@ -1,0 +1,2 @@
+export { NavbarComponent } from './shared/navbar/NavbarComponent';
+export { SidebarComponent } from './shared/sidebar/SidebarComponent';
