+++ src/components/shared/constants.tsx
@@ -1,0 +1,1 @@
+export const drawerWidth = 240;
