+++ src/components/shared/interfaces/AppBarPropsInterface.tsx
@@ -1,0 +1,5 @@
+import { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
+
+export interface AppBarPropsInterface extends MuiAppBarProps {
+  open?: boolean;
+}
