+++ src/components/shared/navbar/NavbarStyle.tsx
@@ -1,0 +1,24 @@
+import MuiAppBar from '@mui/material/AppBar';
+import { drawerWidth } from '../constants';
+import { AppBarPropsInterface } from '../interfaces/AppBarPropsInterface';
+import { styled } from '@mui/material';
+
+export const AppBar = styled(MuiAppBar, {
+  shouldForwardProp: (prop) => prop !== 'open',
+})<AppBarPropsInterface>(({ theme, open }) => ({
+  zIndex: theme.zIndex.drawer + 1,
+  width: `calc(100% - ${theme.spacing(9)})`,
+  boxShadow: 'none',
+  transition: theme.transitions.create(['width', 'margin'], {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  ...(open && {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px - 10px)`,
+    transition: theme.transitions.create(['width', 'margin'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  }),
+}));
