+++ src/pages/main/RootLayout.tsx
@@ -1,30 +1,19 @@
 import { styled, Theme, CSSObject } from '@mui/material/styles';
 import Box from '@mui/material/Box';
 import MuiDrawer from '@mui/material/Drawer';
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
+
 import List from '@mui/material/List';
-import CssBaseline from '@mui/material/CssBaseline';
-import Typography from '@mui/material/Typography';
-import IconButton from '@mui/material/IconButton';
 import ListItem from '@mui/material/ListItem';
 import ListItemButton from '@mui/material/ListItemButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
-import { useState } from 'react';
-import { useLocation } from 'react-router-dom';
-import {
-  Flip,
-  Home,
-  Person,
-  ShoppingBag,
-  SupportAgent,
-} from '@mui/icons-material';
+import { Home, Person, ShoppingBag, SupportAgent } from '@mui/icons-material';
 import { blueTheme } from '../../theme/blueTheme';
 
 import './root.css';
-
-const drawerWidth = 240;
+import { drawerWidth } from '../../components/shared/constants';
+import { useState } from 'react';
+import { NavbarComponent } from '../../components';
 
 const openedMixin = (theme: Theme = blueTheme): CSSObject => ({
   width: drawerWidth,
@@ -55,30 +44,6 @@ const closedMixin = (theme: Theme = blueTheme): CSSObject => ({
   },
 });
 
-interface AppBarProps extends MuiAppBarProps {
-  open?: boolean;
-}
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== 'open',
-})<AppBarProps>(({ theme, open }) => ({
-  zIndex: theme.zIndex.drawer + 1,
-  width: `calc(100% - ${theme.spacing(9)})`,
-  boxShadow: 'none',
-  transition: theme.transitions.create(['width', 'margin'], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px - 10px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
 const Drawer = styled(MuiDrawer, {
   shouldForwardProp: (prop) => prop !== 'open',
 })(({ theme = blueTheme, open }) => ({
@@ -99,49 +64,10 @@ const Drawer = styled(MuiDrawer, {
 
 export const RootLayout = ({ children }: { children: React.ReactNode }) => {
   const [open, setOpen] = useState(false);
-  const handleDrawerToggle = () => {
-    setOpen(!open);
-  };
-  const { pathname } = useLocation();
-
-  const resolvePathname = (pathname: string): string => {
-    switch (pathname) {
-      case '/product' ||
-        '/product/create' ||
-        '/product/edit' ||
-        '/product/delete':
-        return 'Producto';
-      default:
-        return 'Principal';
-    }
-  };
 
   return (
     <Box sx={{ display: 'flex' }}>
-      <CssBaseline />
-      <AppBar position='fixed' open={open} color={'transparent'}>
-        <Toolbar
-          sx={{
-            display: 'flex',
-            justifyContent: 'space-between',
-          }}
-        >
-          <IconButton
-            color='inherit'
-            aria-label='open drawer'
-            onClick={handleDrawerToggle}
-            edge='start'
-            sx={{
-              marginRight: 5,
-            }}
-          >
-            <Flip />
-          </IconButton>
-          <Typography variant='h6' noWrap component='div'>
-            {resolvePathname(pathname)}
-          </Typography>
-        </Toolbar>
-      </AppBar>
+      <NavbarComponent setOpen={setOpen} state={open} />
       <Drawer variant='permanent' open={open}>
         <List>
           {[
